Pre release tweaks
git-svn-id: https://fgrun.svn.sourceforge.net/svnroot/fgrun/trunk/fgrun@330 65451dfc-fe10-0410-907f-9363c9a62196

+++ src/wizard_funcs.cxx
@@ -992,7 +992,7 @@ Wizard::atlas_cb()
 	    }
 	}
 	prefs.set("io-count",(int)io_list.size());
-	for ( i = 0; i < io_list.size(); ++i )
+	for ( i = 0; i < (int)io_list.size(); ++i )
 	{
 	    prefs.set( Fl_Preferences::Name( "io-item-%d", i+1 ), io_list[i].c_str());
 	}
